backend/pyproject.toml
[tool.poetry]
package-mode = false
name = "appointment-system"
version = "0.1.0"
description = "Backend for the nail appointment scheduling system"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.103.0"
uvicorn = "^0.23.2"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.0"}
asyncpg = "^0.28.0"
alembic = "^1.12.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
pydantic = {extras = ["email"], version = "^2.0.0"}
pydantic-settings = "^2.0.0"
python-dotenv = "^1.0.0"
sqllineage = "^1.3.8"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
pytest-html = "^3.2.0"
pytest-cov = "^4.1.0"
httpx = "^0.24.1"
black = "^23.7.0"
flake8 = "^6.1.0"
isort = "^5.12.0"
pre-commit = "^3.3.3"

[tool.poetry.scripts]
start = "src.main.main:start"
dev = "src.main.main:dev"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# Update the pytest configuration to support nested test directories
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
norecursedirs = [".git", "venv", "dist", "build", "__pycache__", "node_modules"]
addopts = "-v"
render_collapsed = "all"


# Enhanced markers for test categories
[tool.pytest.ini_options.markers]
database = "mark tests that require database connection"
auth = "authentication related tests"
api = "API endpoint tests"
slow = "tests that take a long time to run"
unit = "mark tests as unit tests"
integration = "mark tests as integration tests"
e2e = "mark tests as end-to-end tests"

# Keep your existing tool configuration
[tool.black]
line-length = 88
target-version = ["py39"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[tool.ruff]
line-length = 88
target-version = "py312"
select = ["E", "F", "B", "I"]
ignore = []
